be/manual/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4
alembic
psycopg2-binary
pydantic
python-dotenv
httpx
python-telegram-bot  # or use direct HTTP to Telegram API
apscheduler
cryptography
requests
